msdk: Add modifier in gst-inspect
msdkvpp: check if dmabuf is used in set_caps function

We need to know whether to use the dmabuf before creating
the buffer pool, but in fixate_caps function, it is not
possible to determine whether the sinkpad uses the dmabuf.

msdkvpp: Check caps's feature even if there is no peer pad

Even there is no peer pad, it does not mean that the caps
does not support the feature.

msdk: modifier